fix: Support number's positive, negative, nonpositive, nonnegative
At the moment the number mocker fails to take in account the number `0` for minimum and maximum and therefore casts them to `MIN_INTEGER` and `MAX_INTEGER`. The number `0` is used for the following modifiers:

- `number().positive()`
- `number().negative()`
- `number().nonpositive()`
- `number().nonnegative()`

Additionally, `number()._def.minimum.inclusive` and `number()._def.maximum.inclusive` should both be considered when generating the number.

Closes #9